requirements.txt
requests
python-dotenv
openai
rumps
cairosvg
